Render Selector component on level load
This is to ensure that the character button images load as soon as the
level is loaded instead of when the Selector component is rendered the
first time, which caused the images not to be loaded fast enough. The
current solution uses classes to show and hide the Selector component.