Remove hints resize events for small-width devices
Now, there is a screen orientation change event listener for mobile
devices, which triggers hints resizing only when orientation changes.
The previous solution was based on a resize event listener, which
tended to trigger very often on mobile due to browsers showing/hiding
the top bar. The current solution is much better, IMO.